vic-machine: Move extended-help struct to common (#7854)
Move the extended-help struct to the common package, allowing it to
be used by other vic-machine commands.
