+++ cmd/vic-machine/common/help.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package create
+package common
 
 import "gopkg.in/urfave/cli.v1"
