+++ cmd/vic-machine/create/create.go
@@ -60,7 +60,7 @@ type Create struct {
 	memoryReservLimits string
 	cpuReservLimits    string
 
-	help          Help
+	help          common.Help
 	BridgeIPRange string
 
 	Proxies common.Proxies
@@ -640,7 +640,7 @@ func (c *Create) logArguments(op trace.Operation, cliContext *cli.Context) []str
 func (c *Create) Run(clic *cli.Context) (err error) {
 
 	if c.help.AdvancedOptions {
-		cli.HelpPrinter(clic.App.Writer, EntireOptionHelpTemplate, clic.Command)
+		cli.HelpPrinter(clic.App.Writer, common.EntireOptionHelpTemplate, clic.Command)
 		return nil
 	}
